+++ src/lib.rs
@@ -1,7 +1,6 @@
-use colored::*;
 use std::path::PathBuf;
 
-mod cmd;
+mod cli;
 mod dependencies;
 mod error;
 #[cfg(unix)]
@@ -9,141 +8,83 @@ mod execv;
 mod lock;
 mod operations;
 mod paths;
-mod project;
 mod python_info;
 mod run;
 mod settings;
+mod ui;
 #[cfg(windows)]
 mod win_job;
 
-pub use crate::cmd::Command;
-use crate::cmd::SubCommand;
-pub use crate::cmd::{print_error, print_info_1, print_info_2};
-pub use crate::error::*;
-use crate::lock::BumpType;
-use crate::operations::{InitOptions, UpdateOptions};
-pub use crate::paths::{DEV_LOCK_FILENAME, PROD_LOCK_FILENAME};
-use crate::project::{PostInstallAction, ProcessScriptsMode, Project};
+use crate::cli::commands;
+use crate::cli::syntax::SubCommand;
+use crate::paths::{Paths, PathsResolver};
 use crate::python_info::PythonInfo;
+use crate::run::VenvRunner;
+
+pub use crate::cli::syntax::Command;
+pub use crate::error::*;
+pub use crate::paths::{DEV_LOCK_FILENAME, PROD_LOCK_FILENAME};
 pub use crate::settings::Settings;
+pub use crate::ui::{print_error, print_info_1, print_info_2};
 
-fn get_project_path_for_init(cmd: &Command) -> Result<PathBuf, Error> {
-    if let Some(project_path) = &cmd.project_path {
-        Ok(PathBuf::from(project_path))
-    } else {
-        std::env::current_dir().map_err(|e| Error::NoWorkingDirectory { io_error: e })
-    }
+#[derive(Debug)]
+pub struct Metadata {
+    pub dmenv_version: String,
+    pub python_platform: String,
+    pub python_version: String,
 }
 
-fn get_project_path(cmd: &Command) -> Result<PathBuf, Error> {
-    if let Some(project_path) = &cmd.project_path {
-        Ok(PathBuf::from(project_path))
-    } else {
-        look_up_for_project_path()
-    }
+#[derive(Debug)]
+pub enum PostInstallAction {
+    RunSetupPyDevelop,
+    None,
 }
 
-pub fn run(cmd: Command) -> Result<(), Error> {
-    let settings = Settings::from_shell(&cmd);
+#[derive(Debug, Copy, Clone)]
+pub enum ProcessScriptsMode {
+    Safe,
+    Override,
+}
 
-    // Init does not need an existing project
-    if let SubCommand::Init {
-        name,
-        version,
-        author,
-        no_setup_cfg,
-    } = &cmd.sub_cmd
-    {
-        let project_path = get_project_path_for_init(&cmd)?;
-        let mut options = InitOptions::new(name.to_string(), version.to_string());
-        if *no_setup_cfg {
-            options.no_setup_cfg();
-        };
-        if let Some(author) = author {
-            options.author(author);
-        }
-        return operations::init(&project_path, &options);
-    }
+pub enum BumpType {
+    Git,
+    Simple,
+}
 
-    // Run needs additional sanity checks when using `dmenv run`
-    // TODO: try and handle this using StructOpt instead
-    if let SubCommand::Run { ref cmd, .. } = cmd.sub_cmd {
-        if cmd.is_empty() {
-            return Err(new_error(format!(
-                "Missing argument after '{}'",
-                "run".green()
-            )));
-        }
-    }
+#[derive(Default, Debug)]
+/// Represents options passed to `dmenv lock`,
+/// see `cmd::SubCommand::Lock`
+pub struct UpdateLockOptions {
+    pub python_version: Option<String>,
+    pub sys_platform: Option<String>,
+}
 
-    let project_path = get_project_path(&cmd)?;
+#[derive(Debug)]
+pub struct Context {
+    paths: Paths,
+    python_info: PythonInfo,
+    settings: Settings,
+    venv_runner: VenvRunner,
+}
+
+fn get_context(cmd: &Command) -> Result<Context, Error> {
+    let project_path = if let Some(p) = &cmd.project_path {
+        PathBuf::from(p)
+    } else {
+        look_up_for_project_path()?
+    };
     let python_info = PythonInfo::new(&cmd.python_binary)?;
-    let mut project = Project::new(project_path, python_info, settings)?;
-
-    match &cmd.sub_cmd {
-        SubCommand::Install {
-            no_develop,
-            system_site_packages,
-        } => {
-            if *system_site_packages {
-                project.use_system_site_packages()
-            }
-            let post_install_action = if *no_develop {
-                PostInstallAction::None
-            } else {
-                PostInstallAction::RunSetupPyDevelop
-            };
-            project.install(post_install_action)
-        }
-        SubCommand::ProcessScripts { force } => {
-            let mode = if *force {
-                ProcessScriptsMode::Override
-            } else {
-                ProcessScriptsMode::Safe
-            };
-            project.process_scripts(mode)
-        }
-        SubCommand::Clean {} => project.clean_venv(),
-        SubCommand::Develop {} => project.develop(),
-        SubCommand::Lock {
-            python_version,
-            sys_platform,
-            system_site_packages,
-        } => {
-            let update_options = UpdateOptions {
-                python_version: python_version.clone(),
-                sys_platform: sys_platform.clone(),
-            };
-            if *system_site_packages {
-                project.use_system_site_packages();
-            }
-            project.update_lock(update_options)
-        }
-        SubCommand::BumpInLock { name, version, git } => {
-            let bump_type = if *git {
-                BumpType::Git
-            } else {
-                BumpType::Simple
-            };
-            project.bump_in_lock(name, version, bump_type)
-        }
-        SubCommand::Run { ref cmd, no_exec } => {
-            if *no_exec {
-                project.run(&cmd)
-            } else {
-                project.run_and_die(&cmd)
-            }
-        }
-        SubCommand::ShowDeps {} => project.show_deps(),
-        SubCommand::ShowOutDated {} => project.show_outdated(),
-        SubCommand::ShowVenvPath {} => project.show_venv_path(),
-        SubCommand::ShowVenvBin {} => project.show_venv_bin_path(),
-
-        SubCommand::Tidy {} => project.tidy(),
-
-        SubCommand::UpgradePip {} => project.upgrade_pip(),
-        _ => unimplemented!("Subcommand {:?} not handled", cmd.sub_cmd),
-    }
+    let python_version = python_info.version.clone();
+    let settings = Settings::from_shell(&cmd);
+    let paths_resolver = PathsResolver::new(project_path.clone(), python_version, &settings);
+    let paths = paths_resolver.paths()?;
+    let venv_runner = VenvRunner::new(&project_path, &paths.venv);
+    Ok(Context {
+        paths,
+        python_info,
+        settings,
+        venv_runner,
+    })
 }
 
 fn look_up_for_project_path() -> Result<PathBuf, Error> {
@@ -164,6 +105,76 @@ fn look_up_for_project_path() -> Result<PathBuf, Error> {
                 Some(p) => candidate = p.to_path_buf(),
             }
         }
+    }
+}
+
+pub fn run_cmd(cmd: Command) -> Result<(), Error> {
+    let context = get_context(&cmd);
+
+    match &cmd.sub_cmd {
+        SubCommand::Init {
+            name,
+            version,
+            author,
+            no_setup_cfg,
+        } => commands::init(cmd.project_path, name, version, author, !no_setup_cfg),
+
+        SubCommand::Install { no_develop } => {
+            let post_install_action = if *no_develop {
+                PostInstallAction::None
+            } else {
+                PostInstallAction::RunSetupPyDevelop
+            };
+            commands::install(&context?, post_install_action)
+        }
+
+        SubCommand::Clean {} => commands::clean_venv(&context?),
+        SubCommand::Develop {} => commands::develop(&context?),
+        SubCommand::UpgradePip {} => commands::upgrade_pip(&context?),
+
+        SubCommand::ProcessScripts { force } => {
+            let mode = if *force {
+                ProcessScriptsMode::Override
+            } else {
+                ProcessScriptsMode::Safe
+            };
+            commands::process_scripts(&context?, mode)
+        }
+
+        SubCommand::Lock {
+            python_version,
+            sys_platform,
+        } => {
+            let update_options = UpdateLockOptions {
+                python_version: python_version.clone(),
+                sys_platform: sys_platform.clone(),
+            };
+            commands::update_lock(&context?, update_options)
+        }
+
+        SubCommand::BumpInLock { name, version, git } => {
+            let bump_type = if *git {
+                BumpType::Git
+            } else {
+                BumpType::Simple
+            };
+            commands::bump_in_lock(&context?, name, version, bump_type)
+        }
+
+        SubCommand::Run { ref cmd, no_exec } => {
+            if *no_exec {
+                commands::run(&context?, &cmd)
+            } else {
+                commands::run_and_die(&context?, &cmd)
+            }
+        }
+
+        SubCommand::ShowDeps {} => commands::show_deps(&context?),
+        SubCommand::ShowOutDated {} => commands::show_outdated(&context?),
+        SubCommand::ShowVenvPath {} => commands::show_venv_path(&context?),
+        SubCommand::ShowVenvBin {} => commands::show_venv_bin_path(&context?),
+
+        SubCommand::Tidy {} => commands::tidy(&cmd, &context?),
     }
 }