+++ src/operations/venv.rs
@@ -1,10 +1,11 @@
+use colored::*;
 use std::path::{Path, PathBuf};
 
-use crate::cmd::*;
 use crate::error::*;
 use crate::python_info::PythonInfo;
 use crate::run::run;
 use crate::settings::Settings;
+use crate::ui::*;
 
 pub fn clean(venv_path: PathBuf) -> Result<(), Error> {
     print_info_1(&format!("Cleaning {}", venv_path.display()));
@@ -48,6 +49,12 @@ pub fn create(
         args.push("--system-site-packages");
     }
     let python_binary = &python_info.binary;
+    println!(
+        "{} {} {}",
+        "$".blue(),
+        python_binary.display(),
+        args.join(" ")
+    );
     run(&parent_venv_path.to_path_buf(), &python_binary, &args)
 }