+++ src/settings.rs
@@ -1,8 +1,8 @@
-use crate::cmd::Command;
+use crate::cli::syntax::Command;
 
 #[derive(Debug, Clone)]
 /// Represent variables that change behavior of
-/// the Project  or PathsResolver structs.
+/// dmenv commands
 pub struct Settings {
     pub venv_from_stdlib: bool,
     pub venv_outside_project: bool,
@@ -30,6 +30,7 @@ impl Settings {
     pub fn from_shell(cmd: &Command) -> Settings {
         let mut res = Settings {
             production: cmd.production,
+            system_site_packages: cmd.system_site_packages,
             ..Default::default()
         };
         if std::env::var("DMENV_NO_VENV_STDLIB").is_ok() {